+++ packages/client/components/ActivityLibrary/ActivityDetailsSidebar.tsx
@@ -18,9 +18,6 @@ import SendClientSegmentEventMutation from '../../mutations/SendClientSegmentEve
 import StartCheckInMutation from '../../mutations/StartCheckInMutation'
 import StartTeamPromptMutation from '../../mutations/StartTeamPromptMutation'
 import {PALETTE} from '../../styles/paletteV3'
-import SecondaryButton from '../SecondaryButton'
-import GcalModal from '../../modules/userDashboard/components/GcalModal/GcalModal'
-import {useDialogState} from '../../ui/Dialog/useDialogState'
 import {CreateGcalEventInput} from '../../__generated__/StartRetrospectiveMutation.graphql'
 import sortByTier from '../../utils/sortByTier'
 import {MeetingTypeEnum} from '../../__generated__/ActivityDetailsQuery.graphql'
@@ -28,13 +25,14 @@ import {RecurrenceSettings} from '../TeamPrompt/Recurrence/RecurrenceSettings'
 import NewMeetingSettingsToggleAnonymity from '../NewMeetingSettingsToggleAnonymity'
 import NewMeetingSettingsToggleTeamHealth from '../NewMeetingSettingsToggleTeamHealth'
 import NewMeetingSettingsToggleCheckIn from '../NewMeetingSettingsToggleCheckIn'
-import isTeamHealthAvailable from '../../utils/features/isTeamHealthAvailable'
 import StyledError from '../StyledError'
 import FlatPrimaryButton from '../FlatPrimaryButton'
 import NewMeetingActionsCurrentMeetings from '../NewMeetingActionsCurrentMeetings'
 import RaisedButton from '../RaisedButton'
 import NewMeetingTeamPicker from '../NewMeetingTeamPicker'
 import {ActivityDetailsRecurrenceSettings} from './ActivityDetailsRecurrenceSettings'
+import {AdhocTeamMultiSelect, Option} from '../AdhocTeamMultiSelect/AdhocTeamMultiSelect'
+import ScheduleMeetingButton from './ScheduleMeetingButton'
 
 interface Props {
   selectedTemplateRef: ActivityDetailsSidebar_template$key
@@ -67,11 +65,12 @@ const ActivityDetailsSidebar = (props: Props) => {
         featureFlags {
           gcal
         }
+        ...AdhocTeamMultiSelect_viewer
+        ...ScheduleMeetingButton_viewer
       }
     `,
     viewerRef
   )
-  const hasGcalFlag = viewer.featureFlags.gcal
 
   const teams = useFragment(
     graphql`
@@ -95,10 +94,11 @@ const ActivityDetailsSidebar = (props: Props) => {
         actionSettings: meetingSettings(meetingType: action) {
           ...NewMeetingSettingsToggleCheckIn_settings
         }
+        ...NewMeetingSettingsToggleTeamHealth_team
         ...NewMeetingTeamPicker_selectedTeam
         ...NewMeetingTeamPicker_teams
         ...NewMeetingActionsCurrentMeetings_team
-        ...GcalModal_team
+        ...ScheduleMeetingButton_team
       }
     `,
     teamsRef
@@ -126,13 +126,19 @@ const ActivityDetailsSidebar = (props: Props) => {
       templateTeam ??
       sortByTier(availableTeams)[0]!
   )
-  const {onError, onCompleted, submitting, submitMutation, error} = useMutationProps()
+  const mutationProps = useMutationProps()
+  const {onError, onCompleted, submitting, submitMutation, error} = mutationProps
   const history = useHistory()
-  const {
-    isOpen: isScheduleDialogOpen,
-    open: openScheduleDialog,
-    close: closeScheduleDialog
-  } = useDialogState()
+
+  const [selectedUsers, setSelectedUsers] = React.useState<Option[]>([])
+
+  const oneOnOneTeamInput = selectedUsers[0]
+    ? {
+        email: selectedUsers[0].email,
+        // TODO: always pick default orgId for now, this will be changed in next PRs
+        orgId: selectedTeam.orgId
+      }
+    : null
 
   const handleStartActivity = (gcalInput?: CreateGcalEventInput) => {
     if (submitting) return
@@ -153,7 +159,7 @@ const ActivityDetailsSidebar = (props: Props) => {
     } else if (type === 'action') {
       StartCheckInMutation(
         atmosphere,
-        {teamId: selectedTeam.id, gcalInput},
+        {teamId: !oneOnOneTeamInput ? selectedTeam.id : null, oneOnOneTeamInput, gcalInput},
         {history, onError, onCompleted}
       )
     } else {
@@ -234,17 +240,17 @@ const ActivityDetailsSidebar = (props: Props) => {
 
         <div className='flex grow flex-col gap-2'>
           {selectedTemplate.id === 'oneOnOneAction' ? (
-            // TODO: replace it with new user picker
-            <NewMeetingTeamPicker
-              positionOverride={MenuPosition.UPPER_LEFT}
-              onSelectTeam={(teamId) => {
-                const newTeam = availableTeams.find((team) => team.id === teamId)
-                newTeam && setSelectedTeam(newTeam)
-              }}
-              selectedTeamRef={selectedTeam}
-              teamsRef={availableTeams}
-              customPortal={teamScopePopover}
-            />
+            <div className='rounded-lg bg-slate-200 p-3'>
+              <div className='text-gray-700 pb-3 text-lg font-semibold'>Teammate</div>
+              <AdhocTeamMultiSelect
+                viewerRef={viewer}
+                onChange={(newUsers: Option[]) => {
+                  setSelectedUsers(newUsers)
+                }}
+                value={selectedUsers}
+                multiple={false}
+              />
+            </div>
           ) : (
             <NewMeetingTeamPicker
               positionOverride={MenuPosition.UPPER_LEFT}
@@ -277,9 +283,10 @@ const ActivityDetailsSidebar = (props: Props) => {
               {type === 'retrospective' && (
                 <>
                   <NewMeetingSettingsToggleCheckIn settingsRef={selectedTeam.retroSettings} />
-                  {isTeamHealthAvailable(selectedTeam.tier) && (
-                    <NewMeetingSettingsToggleTeamHealth settingsRef={selectedTeam.retroSettings} />
-                  )}
+                  <NewMeetingSettingsToggleTeamHealth
+                    settingsRef={selectedTeam.retroSettings}
+                    teamRef={selectedTeam}
+                  />
                   <NewMeetingSettingsToggleAnonymity settingsRef={selectedTeam.retroSettings} />
                 </>
               )}
@@ -298,15 +305,12 @@ const ActivityDetailsSidebar = (props: Props) => {
               <div className='flex grow flex-col justify-end gap-2'>
                 {error && <StyledError>{error.message}</StyledError>}
                 <NewMeetingActionsCurrentMeetings team={selectedTeam} />
-                {hasGcalFlag && (
-                  <SecondaryButton
-                    onClick={openScheduleDialog}
-                    waiting={submitting}
-                    className='h-14'
-                  >
-                    <div className='text-lg'>Schedule</div>
-                  </SecondaryButton>
-                )}
+                <ScheduleMeetingButton
+                  handleStartActivity={handleStartActivity}
+                  mutationProps={mutationProps}
+                  teamRef={selectedTeam}
+                  viewerRef={viewer}
+                />
                 <FlatPrimaryButton
                   onClick={() => handleStartActivity()}
                   waiting={submitting}
@@ -319,12 +323,6 @@ const ActivityDetailsSidebar = (props: Props) => {
           )}
         </div>
       </div>
-      <GcalModal
-        closeModal={closeScheduleDialog}
-        isOpen={isScheduleDialogOpen}
-        handleStartActivityWithGcalEvent={handleStartActivity}
-        teamRef={selectedTeam}
-      />
     </>
   )
 }