+++ packages/server/graphql/mutations/startRetrospective.ts
@@ -94,7 +94,7 @@ export default {
     const template = await dataLoader.get('meetingTemplates').load(selectedTemplateId)
     await Promise.all([
       r.table('NewMeeting').insert(meeting).run(),
-      updateMeetingTemplateLastUsedAt(selectedTemplateId)
+      updateMeetingTemplateLastUsedAt(selectedTemplateId, teamId)
     ])
 
     // Disallow accidental starts (2 meetings within 2 seconds)
@@ -121,10 +121,10 @@ export default {
         .run(),
       updateTeamByTeamId(updates, teamId)
     ])
-    createGcalEvent({gcalInput, meetingId, teamId, viewerId, dataLoader})
     IntegrationNotifier.startMeeting(dataLoader, meetingId, teamId)
     analytics.meetingStarted(viewerId, meeting, template)
-    const data = {teamId, meetingId}
+    const {error} = await createGcalEvent({gcalInput, meetingId, teamId, viewerId, dataLoader})
+    const data = {teamId, meetingId, hasGcalError: !!error?.message}
     publish(SubscriptionChannel.TEAM, teamId, 'StartRetrospectiveSuccess', data, subOptions)
     return data
   }