+++ packages/server/graphql/types/User.ts
@@ -49,7 +49,6 @@ import {TimelineEventConnection} from './TimelineEvent'
 import TimelineEventTypeEnum from './TimelineEventTypeEnum'
 import TimelineEvent from '../../database/types/TimelineEvent'
 import {RDatum} from '../../database/stricterR'
-import {getAccessibleTeamIdsForUser} from '../getAccessibleTeamIdsForUser'
 
 const User: GraphQLObjectType<any, GQLContext> = new GraphQLObjectType<any, GQLContext>({
   name: 'User',
@@ -240,7 +239,11 @@ const User: GraphQLObjectType<any, GQLContext> = new GraphQLObjectType<any, GQLC
             edges: []
           }
         }
-        const validTeamIds = await getAccessibleTeamIdsForUser(viewerId, teamIds, dataLoader)
+        const userTeamMembers = await dataLoader.get('teamMembersByUserId').load(viewerId)
+        const accessibleTeamIds = userTeamMembers.map(({teamId}) => teamId)
+        const validTeamIds = teamIds
+          ? teamIds.filter((teamId: string) => accessibleTeamIds.includes(teamId))
+          : accessibleTeamIds
 
         if (viewerId !== id && !isSuperUser(authToken)) return null
         const dbAfter = after ? new Date(after) : r.maxval
@@ -595,17 +598,28 @@ const User: GraphQLObjectType<any, GQLContext> = new GraphQLObjectType<any, GQLC
     teams: {
       type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(Team))),
       description: 'all the teams the user is on that the viewer can see.',
+      args: {
+        includeArchived: {
+          type: GraphQLBoolean,
+          defaultValue: false,
+          description:
+            'If true, returns archived teams as well; otherwise only return active teams. Default to false.'
+        }
+      },
       resolve: async (
         {id: userId}: {id: string},
-        _args: unknown,
+        {includeArchived},
         {authToken, dataLoader}: GQLContext
       ) => {
         const viewerId = getUserId(authToken)
         const user = (await dataLoader.get('users').load(userId))!
-        const teamIds =
+        const activeTeamIds =
           viewerId === userId || isSuperUser(authToken)
             ? user.tms
             : user.tms.filter((teamId: string) => authToken.tms.includes(teamId))
+        const teamIds = includeArchived
+          ? (await dataLoader.get('teamMembersByUserId').load(userId)).map(({teamId}) => teamId)
+          : activeTeamIds
         const teams = (await dataLoader.get('teams').loadMany(teamIds)).filter(isValid)
         teams.sort((a, b) => (a.name > b.name ? 1 : -1))
         return teams